spring-ws-support should be a fragment bundle.
This bundle adds extra classes to spring-ws-core and exports some of the
same packages. It should attach to that bundle and complement it.

If deployed as it currently is, it will result in use constraint
violations for importing bundles since packages like
org.springframework.ws.transport will resolve via two dependency chains.

The last upstream version with OSGi metadata also packaged this bundle
as a fragment.

https://repo1.maven.org/maven2/org/springframework/ws/spring-ws-support/2.1.4.RELEASE/spring-ws-support-2.1.4.RELEASE.jar